study/20251121_rag_ch14_17/langchain-practice/pyproject.toml
[project]
name = "langchain-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "arxiv==2.1.3",
    "chromadb==0.6.3",
    "deepl==1.21.1",
    "docx2txt==0.8",
    "faiss-cpu==1.10.0",
    "flashrank==0.2.10",
    "google-search-results==2.4.2",
    "grandalf==0.8",
    "huggingface-hub==0.29.3",
    "jupyter==1.1.1",
    "kiwipiepy==0.20.4",
    "konlpy==0.6.0",
    "langchain==0.3.21",
    "langchain-anthropic==0.3.10",
    "langchain-azure-ai==0.1.2",
    "langchain-chroma==0.2.2",
    "langchain-cohere==0.4.3",
    "langchain-community==0.3.20",
    "langchain-core==0.3.46",
    "langchain-elasticsearch==0.3.2",
    "langchain-experimental==0.3.4",
    "langchain-google-genai==2.1.0",
    "langchain-huggingface==0.1.2",
    "langchain-milvus==0.1.8",
    "langchain-ollama==0.2.3",
    "langchain-openai==0.3.9",
    "langchain-teddynote==0.3.44",
    "langchain-text-splitters==0.3.7",
    "langchainhub==0.1.21",
    "langgraph==0.3.18",
    "langsmith==0.3.18",
    "lark==1.2.2",
    "llama-index-core==0.11.23",
    "llama-index-readers-file==0.2.2",
    "llama-parse==0.6.4.post1",
    "lxml==5.3.1",
    "matplotlib==3.10.1",
    "mypy==1.15.0",
    "nest-asyncio==1.6.0",
    "notebook==7.3.3",
    "open-clip-torch==2.31.0",
    "openai==1.67.0",
    "pandas==2.2.3",
    "pdfminer-six==20231228",
    "pdfplumber==0.11.5",
    "pillow==10.4.0",
    "pinecone==5.4.2",
    "protobuf==3.20.3",
    "pydantic==2.10.6",
    "pymilvus==2.5.5",
    "pymupdf==1.25.4",
    "pymupdf4llm==0.0.17",
    "pypdf==4.3.1",
    "python-dotenv==1.0.1",
    "ragas==0.2.14",
    "rank-bm25==0.2.2",
    "rapidocr-onnxruntime==1.4.4",
    "redis==5.2.1",
    "rouge-score==0.1.2",
    "scikit-learn==1.6.1",
    "seaborn==0.13.2",
    "sqlalchemy==2.0.39",
    "streamlit==1.43.2",
    "tenacity==8.5.0",
    "tiktoken==0.9.0",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "unstructured==0.17.0",
    "wikipedia==1.4.0",
]
